requirements.txt
jira>=3.5.0
google-auth>=2.23.0
google-api-python-client>=2.100.0
